Setting up to handle pdfs
